refactor(binder): split flow statement binding
